Add files via upload
This is a smart contract wallet which has one owner and is able to receive funds (in ether, wei, gwei etc) in all situations. 
Further functionalities include the owner being able to send funds to any kind of address, whether it be another smart contract address or an externally owned account (EOA) with a private key. 
Moreover, it allows only the owner to assign allowance and permission to EOAs or smart contracts to spend from the contract. 
In case private key is lost and hence ether are lost; it is possible to set a new owner to a different address, than the initially lost one, by the voting requirement of minimum 3 out of 5 guardians voting for a new owner-assuming guardians addresses are pre-set by owner (lost account). (Only a single vote can be counted by a single guardian; a mapping within a mapping is used to avoid double voting) 

This is my first solidity project after spending hours learning the basics! 
Do suggest what I should try next...  ERC20 token maybe?