Add files via upload
This project allows a wallet address to interact through a purchaser contract which enables them to participate in the lottery system by adding their amount into the contract and further purchasing tickets from the lottery contract. The participant can access functions from the lottery contract through an interface. Only the participant has the ability to send and call the lottery contract externally through the purchaser contract. Only the owner of the lottery contract can distribute the prize to the participants. The award is 80% of the entire lottery pool and is sent to the winners deployed 'purchaser' contract from which they can then withdraw the prize. Feel free to add/remove any functionality into the contracts, I have also created the purchaser contract for demo purposes.