Add files via upload
This project was just for fun and to test functionalities, through simple structs and mappings, on solidity. A system which assigns an ID to the student on making their entry; The ID is unique and cannot be duplicated for another student. Upon removal of student, a simple swapping array technique is used using 'for' loop and an 'if' condition to find and .pop() student. Note that only the deployer/owner/admin can edit/update the data.  Enum functionality is used to edit specific field respective to student and so on. Feel free to add/remove code from the .sol file where you think improvement is needed. (dont forget to let me know on my email thanks!)